requirements.txt
Django~=4.1.2
djangorestframework
pygments
httpie
drf-spectacular

# docker run -p 80:8080 -e SWAGGER_JSON=/schema.yml -v ${PWD}/schema.yml:/schema.yml swaggerapi/swagger-ui
# python.exe .\manage.py spectacular --color --file schema.yml
